requirements.txt
streamlit>=1.35.0
openai>=1.30.0
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.0
pytest>=8.0.0
openpyxl>=3.1.0
supabase>=2.0.0
